backend/requirements.txt
# Core FastAPI & Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0

# Database
SQLAlchemy==2.0.35
# Optional: PostgreSQL driver. Commented out to avoid build issues on Windows/Python 3.13.
# Enable only when using a Postgres DATABASE_URL.
# psycopg2-binary==2.9.9
alembic==1.13.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Data Processing
pandas==2.2.2
numpy==2.0.2
openpyxl==3.1.2

# OCR & Computer Vision
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.0.1
easyocr==1.7.0

# AI/ML Models
scikit-learn==1.5.2
transformers==4.35.2
torch==2.1.1
prophet==1.1.5
xgboost==2.1.1
google-generativeai==0.7.2
nltk==3.8.1

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78

# Background Tasks
celery==5.4.0
redis==5.0.8

# HTTP & API
httpx==0.27.2
requests==2.31.0

# Financial & GST
python-gstin==0.1.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
